% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_arma.R
\name{grouped_arma}
\alias{grouped_arma}
\title{Fit grouped ARMA model to time series}
\usage{
grouped_arma(x, order = c(1, 1), lag = NULL, coef = NULL,
  include.intercept = TRUE, series = NULL, qr.tol = 1e-07, ...)
}
\arguments{
\item{x}{A numeric vector or time series.}
}
\value{
Object of time \code{arma}.
}
\description{
Fit an grouped ARMA model to a univariate time series by conditional least squares.
}

